Consolidate Card and CardGroup
