+++ src/components/ComponentFactory.js
@@ -4,8 +4,8 @@ import Admonition, { admonitionMap } from './Admonition';
 import Banner from './Banner/Banner';
 import BlockQuote from './BlockQuote';
 import Button from './Button';
-import Card from './Card';
-import CardGroup from './CardGroup';
+import Card from './Card/Card';
+import CardGroup from './Card/CardGroup';
 import Chapter from './Chapters/Chapter';
 import Chapters from './Chapters/Chapters';
 import Code from './Code/Code';
