+++ tests/unit/Card.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import Card from '../../src/components/Card';
-import { theme } from '../../src/theme/docsTheme';
 import { ThemeProvider } from '@emotion/react';
+import Card from '../../src/components/Card/Card';
+import { theme } from '../../src/theme/docsTheme';
 // data for this component
 import mockData from './data/Card.test.json';
