+++ tests/unit/CardGroup.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import CardGroup from '../../src/components/CardGroup';
-import { theme } from '../../src/theme/docsTheme';
 import { ThemeProvider } from '@emotion/react';
+import CardGroup from '../../src/components/Card/CardGroup';
+import { theme } from '../../src/theme/docsTheme';
 // data for this component
 import mockData from './data/CardGroup.test.json';
